Add hotkey "R" to respect symmetries, rather than always respecting them.
And update the docs.

For anything above six armies, it feels like the symmetry is hurting
(by causing stripes to "stick" on the main diagonal, and perhaps locking
coincidentally symmetrical stripes together when they shouldn't be locked)
more than it helps (it doesn't help).

So, make it an opt-in feature via hotkey "R", and document the other
hotkeys I've added recently.